requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-CORS
Flask-Login
mysql-connector-python
PyMySQL
python-dotenv
numpy
dotenv
requests
gunicorn
apscheduler